+++ src/app/grapkimbo/Components/VisualOutline.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "commons.h"
+#include "engine/commons.h"
 #include <aunteater/Component.h>
 
 
@@ -10,7 +12,7 @@ namespace grapito
 
 struct VisualOutline : public aunteater::Component<VisualOutline>
 {
-    VisualOutline(math::sdr::Rgb aColor, double aOvershoot) :
+    VisualOutline(math::sdr::Rgb aColor, float aOvershoot) :
         color{aColor},
         overshoot{aOvershoot, aOvershoot}
     {}
